Fulfill clippy suggestions
