+++ src/lib.rs
@@ -1,4 +1,6 @@
 pub mod cipher;
 pub mod decrypter;
+pub mod encrypter;
 mod eme;
 pub mod obscure;
+pub mod constants;
